requirements.txt
streamlit>=1.35.0
llama-index-core>=0.10.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-readers-file>=0.1.0
chromadb>=0.5.0
pypdf>=4.0.0
python-docx>=1.1.0
pytest>=8.0.0
openai>=1.3.0
python-dotenv>=1.0.0
